[cppgc] Move PageAllocator To CppgcPlatformAdapter
'CppHeap' has a 'CppgcPlatformAdapter' class which it sets as the
platform for 'HeapBase'.  It later expects to be able to call
'HeapBase::platform()', which returns a 'cppgc::Platform', and
'static_cast' it back to 'CppgcPlatformAdapter'.

Previously, 'HeapBase' was creating a 'PlatformWithPageAllocator' class,
which internally fetches the global page allocator, if the incoming
'cppgc::Platform::GetPageAllocator()' method returned 'nullptr'.  The
'HeapBase::platform()' method then returned this
'PlatformWithPageAllocator' instance, which isn't 'static_cast'-able to
'CppgcPlatformAdapter'.

This commit merges 'PlatformWithPageAllocator' into
'CppgcPlatformAdapter', so that the 'static_cast' remains valid when a
platform without a 'PageAllocator' is provided.